Separate employee checking behaviour
